+++ test/beast/core/_detail_base64.cpp
@@ -11,6 +11,7 @@
 #include <boost/beast/core/detail/base64.hpp>
 
 #include <boost/beast/_experimental/unit_test/suite.hpp>
+#include <string>
 
 namespace boost {
 namespace beast {
@@ -19,6 +20,35 @@ namespace detail {
 class base64_test : public beast::unit_test::suite
 {
 public:
+    std::string
+    base64_encode(
+        std::uint8_t const* data,
+        std::size_t len)
+    {
+        std::string dest;
+        dest.resize(base64::encoded_size(len));
+        dest.resize(base64::encode(&dest[0], data, len));
+        return dest;
+    }
+
+    std::string
+    base64_encode(string_view s)
+    {
+        return base64_encode (reinterpret_cast <
+            std::uint8_t const*> (s.data()), s.size());
+    }
+
+    std::string
+    base64_decode(string_view data)
+    {
+        std::string dest;
+        dest.resize(base64::decoded_size(data.size()));
+        auto const result = base64::decode(
+            &dest[0], data.data(), data.size());
+        dest.resize(result.first);
+        return dest;
+    }
+
     void
     check (std::string const& in, std::string const& out)
     {